src/main/webapp/app/entities/1microSampleApplication/table-8/delete/table-8-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { Table8Service } from '../service/table-8.service';

import { Table8DeleteDialogComponent } from './table-8-delete-dialog.component';

describe('Table8 Management Delete Component', () => {
  let comp: Table8DeleteDialogComponent;
  let fixture: ComponentFixture<Table8DeleteDialogComponent>;
  let service: Table8Service;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, Table8DeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(Table8DeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(Table8DeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(Table8Service);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
